pyproject.toml
[tool.poetry]
name = "aola"
version = "0.1.0"
description = ""
authors = ["alex"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.1"
django = "^4.1.7"
djangorestframework = "^3.14.0"
python-environ = "^0.4.54"
pillow = "^9.4.0"
psycopg2-binary = "^2.9.5"
django-polymorphic = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
